Update and rename SalaryCalcTest.java to src/test/java/SalaryCalc/SalaryCalcTest.java
